PAAC-154: Moved TaperedAllowanceCalculator to calculators.internal. Moved and renamed calculators.results.SummaryResultCalculator.scala to calculators.internal.BasicAllowanceCalculator.scala
