+++ app/calculators/internal/PeriodCalculator.scala
@@ -17,8 +17,6 @@
 package calculators.internal
 
 import models._
-import calculators.ExtendedSummaryCalculator
-import calculators.results.SummaryResultCalculator
 import calculators.internal.Utilities._
 import calculators.Utilities._
 
@@ -43,6 +41,6 @@ object PeriodCalculator {
     else if (contribution.isPeriod2)
       new Period2Calculator
     else
-      new SummaryResultCalculator(annualAllowanceInPounds, previousPeriods, contribution) with PeriodCalculator
+      new BasicAllowanceCalculator(annualAllowanceInPounds, previousPeriods, contribution) with PeriodCalculator
   }
 }
